renamed and refactored InputValidator.Class -- not even sure if we'll actually need it come JavaScript but we'll see.
