+++ src/test/java/com/revature/projectzero/services/UserServiceTestSuite.java
@@ -2,7 +2,6 @@ package com.revature.projectzero.services;
 
 import com.revature.projectzero.documents.AppUser;
 import com.revature.projectzero.repositories.UserRepository;
-import com.revature.projectzero.util.InputValidator;
 import com.revature.projectzero.util.UserSession;
 import com.revature.projectzero.util.exceptions.AuthenticationException;
 import com.revature.projectzero.util.exceptions.InvalidEntryException;
@@ -23,7 +22,7 @@ public class UserServiceTestSuite {
     // Mock dependencies required for the system under test
     private UserRepository mockUserRepo;
     private UserSession mockSession;
-    private InputValidator mockValidator;
+    private InputValidatorService mockValidator;
 
     // Initialize objects before testing
     @Before
@@ -31,7 +30,7 @@ public class UserServiceTestSuite {
 
         mockUserRepo = mock(UserRepository.class);
         mockSession = mock(UserSession.class);
-        mockValidator = mock(InputValidator.class);
+        mockValidator = mock(InputValidatorService.class);
         sut = new UserService( mockUserRepo, mockSession, mockValidator);
 
     }
